Fix legend for cross section plots
The legend for the cross section process type plots was mislabeled in the
case of PP and NP collisions. The applied categorization, where any
inelastic 2->2 process is parametrized is correct in the case of
kaon-nucleon interactions, but wrong in the case of PP or NP. Here, we
actually form resonances and the cross sections are not parametrized. A
splitting is necessary. Further info can be found in the comment of the
modified file.