Increase box size of detailed balance thermal box
And fix ticks for the y-axis of the lowest panel detailed balance plots.
The ticks were too close, such that the labels would overlap and one
could not read anything anymore.
This also required moving the dummy entries for the legend slightly
further up to make sure they do not show up in the frame.

Fixes #6