Minor bugfixes
1) If the mT or pT spectra are scaled logarithmically on the y-axis for
a certain energy but the corresponding data for an energy that is plotted afterwards is basically zero (can be the case for rare particles), the plotting crashes. This is because the zero datapoints cannot be plotted on the logarithmic scale. We thus need to rescale it to be linear.

2) For the corss sections, the data files that are read in for the
plotting are huge, so that reading the same file multiple times at the
same time to generate the plots on multiple cores does not work from
time to time. Building the plots on only one core solves the issue.