Finish stochastic elastic box plotting
Now, also the comparison to previous version will be plotted for both
geometric and stochastic criterion. Plus, the plot style was changed to
have 2 seperates subplots.

Note: This breaks the plotting to previous version 1.5, but, since the
old 1.6 results will soon be the previous version, this will only be the
case for a breif time.