Update runtime and equilibration time for N_pi_deutron detailed balance box
Between SMASH-1.8 and SMASH-2.0, the cross section cutoff was modified
such that the deutron cross sections are automatically cut at 2000 mb,
insetad of the default of 200 mb. This results in larger equilibration
times for the N_pi_deutron box. Hence, the runtime for this box is
increased to 6000 fm and the estimated equilibration time to 4000 fm.