Adjust grid cell length for the stochastic criterion (again)
The elastic_box_scatrate_vs_dt_sims target was failing for the 
previously small grid length. The grid length is therefore adjusted 
again to be at bit larger.

The reason why previously (when the same grid cell length determination 
was used for stochastic and geometric) the target was running, is that 
the grid cell length is also timestep dependend. Usually the dependence 
is neglectable, but for the elastic_box_scatrate_vs_dt_sims target it is 
sizable (cf. smash function compute_min_cell_length() in epxeriment.h).