Avoid crashes when reading unfinished files
It often happens that SMASH produces unfinished runs, for example because
the time ended on the cluster, or SMASH crashed for some reason. Often
information from these unfinished runs is very useful, so it makes sense
for the read-out script to consume all the useful information and stop
at the right moment without crash.

This little modification allows the read-in script to read input file
until the first broken block without crashing.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>