Fixed bug in read_binary_output.py
In read_binary_output.py, after reading a block, we check if the header (p,f or i) of the next block is valid to prevent reading corrupted files, ee.g. because of a system failure or a cluster job time out. However, the case of an interaction 'i' was not handled properly and this commit hofully fixes the bug.